Added control groups for child distance and border width (and border width matches edge width)
